fixing KeyboardInterrupt signal from xboard + Refactoring
