+++ .gitattributes
@@ -1,0 +1,3 @@
+*.tar filter=lfs diff=lfs merge=lfs -text
+core/models/*.tar filter=lfs diff=lfs merge=lfs -text
+*.debug filter=lfs diff=lfs merge=lfs -text
