+++ play.py
@@ -1,17 +1,16 @@
 import chess
-from coach.trainer import Coach
-from engine import MODEL
-from main import UmkaEvaluator
-from minimax import MiniMax
-from utils import show_board
+from chess.pgn import Game
+from core.minimax import MiniMax
+from core.umka import Umka
 
-coach = Coach(MODEL)
-evaluator = UmkaEvaluator(coach.get_model())
-brain = MiniMax(evaluator)
+umka = Umka(path="model/model.pth.tar", training_enabled=False)
+brain = MiniMax(umka)
 board = chess.Board()
 
 
 while not board.is_game_over():
-    show_board(board)
     move = brain.run(board, depth=2)
     board.push(move)
+
+game = Game().from_board(board)
+print(game)
