+++ settings.py
@@ -1,2 +1,5 @@
 import os
 PROJECT_ROOT_DIR = os.path.dirname(__file__)
+MODELS_DIR = os.path.join(PROJECT_ROOT_DIR, "core", "models")
+
+DEPTH = 3
